refactor(package-manager): split the optimistic repeat-install check
`optimistic_repeat_install.rs` decided one question — may this install
be skipped — but carried every probe that feeds the answer in the same
file. The probes move out by subject:

- `conflict_markers` — refusing when a lockfile still has merge markers
- `local_file_deps` — specs pointing at the local filesystem
- `timestamps` — mtimes, and the baseline a recorded validation uses
- `manifest_agreement` — whether the manifests still match the lockfile
- `settings` — drift between recorded and current settings
- `deps_status` — the pre-run check behind `verifyDepsBeforeRun`

The decision itself stays, at 625 lines rather than 1899, and now reads
as the sequence of questions it asks.

Re-exports are explicit and split by each item's existing visibility.
Tests import from the module that owns what they exercise.

Pure code motion: no behaviour change, verified by comparing every
content line of the original against the union of the new modules.
